Use same token params for web and API
