+++ app/controllers/api/v0/tokens_controller.rb
@@ -1,10 +1,10 @@
 class Api::V0::TokensController < ApiController
+  include AuthParams
+
   def create
-    body = request.body.read
-      authentication = JSON.parse(body)
-      response = CDX::Authenticator.new(authentication).perform
-      session[:system_session_token] = response[:token]
-      response[:token] = session.id
-      render json: response.to_json
+    response = CDX::Authenticator.new(auth_params).perform
+    session[:system_session_token] = response[:token]
+    response[:token] = session.id
+    render json: response.to_json
   end
 end
