+++ app/controllers/concerns/auth_params.rb
@@ -1,0 +1,5 @@
+module AuthParams
+  def auth_params
+    params.fetch(:token, {}).permit(:user_id, :password)
+  end
+end
