+++ db/structure.sql
@@ -136,6 +136,8 @@ INSERT INTO schema_migrations (version) VALUES ('20150924181500');
 
 INSERT INTO schema_migrations (version) VALUES ('20160112190459');
 
+INSERT INTO schema_migrations (version) VALUES ('20160115181255');
+
 INSERT INTO schema_migrations (version) VALUES ('20160115203719');
 
 INSERT INTO schema_migrations (version) VALUES ('20160115204128');
